move dataset validation code to separate module
main dataset module was getting too big
